requirements.txt
blinker==1.7.0
boto3==1.34.84
botocore==1.34.84
click==8.1.7
Flask==2.3.3
Flask-DotEnv==0.1.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
piexif==1.1.3
pillow==10.3.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==2.3.8
WTForms==3.1.2
